refactor-dog - Transform getMyDogDetails into getDogProfile
